// frontend/smartiphr/src/app/models/indumentiIngresso.ts
export class IndumentiIngresso {
  indumento: string;
  nome: string;
  quantita: number;
  dataCaricamento?: Date;
  conforme?: Boolean;
  noninelenco?: Boolean;
  richiesto?: Boolean;
  operatore?: string;
  nomeOperatore?: string;
  paziente?: string;
  note?: string;
  nomePaziente?: string;
  _id?: string;
}